fs: move and fix generic permission checks
As defined by POSIX permissions are checked for the first matching class,
user, group or others.
Only the permissions of this first matching class are considered.

For example:
Consider a process running for user and group.
A file with user:group -w- rw- --- is not readable by user, and
thus not readable by the process.
However, it is readable by anyone else in the group.

Additionally, since the permission checks are generic and not exclusive
to the ext2 driver, they are moved to the vfs code.