do not open files during path resolution
Opening the current path component was done because the opened
file could be used for a call to vfs_stat as well as vfs_readlink.

However, a path may contain traversable directories which can not
be opened with vfs_open because they are neither readable nor writeable.

The following path /traversable/foo can not be resolved by uid 1000, if
the traversable directory has the following permissions:

	d--x--x--x 1000 1000 traversable

This is fixed by using vfs_fstat and the current absolute path to
retrieve the attributes of the current path component.