process: fix interpreter support
Previously, the argv array was not altered. However, this does not
work when the script was executed using execp.
When executing an executable from the path the interpreter has
no way to determine what file it should load.

Now we do the correct thing and add the path to the executed file
as second argument shifting everything else to the right.

TODO: Pass the path to the interpreter as first argument.